Cleaned up and added search bar basics
I am not sure how any of this happened but i got it to look okay.  Still not sure if we should get rid of the boarder around the screen.  Added a simple implimentation of the search bar but was having trouble with it's submit button.  Also, some of the css is conflicting so I need to learn how to change some of it to only effect things with certian id's.